Split board features into dedicated directories.
