--- lib/board_adapters/board.hpp
@@ -1,7 +1,0 @@
-/** \file
- *
- * This is a \ref proxy_header.
- */
-
-#pragma once
-#include <board_interface.hpp>
