+++ lib/board_adapters/input_device_interface/input_devices.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include <board_interface.hpp>
+
+namespace board
+{
+void setup_input_devices(const HmiHandler callbackFunction);
+}
