+++ lib/board_adapters/sound_output_interface/sound_output.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace board
+{
+void setup_sound();
+}
