+++ lib/board_adapters/status_indicators/status_indicators.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace board
+{
+void setup_status_indicators();
+}
